+++ packages/terra-application/src/disclosure-manager/index.js
@@ -1,6 +1,17 @@
 export { default } from 'terra-disclosure-manager';
 export {
-  withDisclosureManager, disclosureManagerShape, DisclosureManagerContext, DisclosureManagerDelegate, DisclosureManagerHeaderAdapterContext, DisclosureManagerHeaderAdapter, availableDisclosureSizes, availableDisclosureHeights, availableDisclosureWidths,
+  withDisclosureManager,
+  disclosureManagerShape,
+  DisclosureManagerContext,
+  DisclosureManagerDelegate,
+  DisclosureManagerHeaderAdapterContext,
+  DisclosureManagerHeaderAdapter,
+  availableDisclosureSizes,
+  availableDisclosureHeights,
+  availableDisclosureWidths,
 } from 'terra-disclosure-manager';
 
-export { closeMostRecentDisclosure, getActiveDisclosureCount } from './_disclosureCallbacks';
+export {
+  closeMostRecentDisclosure,
+  getActiveDisclosureCount,
+} from './_disclosureCallbacks';